refactor put steps classes in separate folder steps
