+++ src/steps/BaseStep.ts
@@ -1,6 +1,6 @@
-import { BuilderData } from './builder'
-import { Condition, Expression, PostgresBinder } from './models'
-import { BooleanColumn } from './columns'
+import { BuilderData } from '../builder'
+import { Condition, Expression, PostgresBinder } from '../models'
+import { BooleanColumn } from '../columns'
 import { LogicalOperator, Parenthesis } from './steps'
 
 export class BaseStep {
